# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Learning Resources API",
        "description": "API documentation for the Learning Resources React Frontend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/api/resources": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Get list of all learning resources",
                "operationId": "0a6700b47dcce2d2d071b3000f20aa60",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Advanced Calculus Video"
                                            },
                                            "reviews_avg_rating": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 4.5
                                            },
                                            "ai_summary": {
                                                "type": "string",
                                                "example": "Most users found this highly comprehensive."
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Resources"
                ],
                "summary": "Submit a new learning resource (YouTube or Textbook ISBN)",
                "operationId": "e889331b6f6c163c83371a691ee6b787",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "identifier"
                                ],
                                "properties": {
                                    "type": {
                                        "description": "Must be 'youtube' or 'textbook'",
                                        "type": "string",
                                        "example": "textbook"
                                    },
                                    "identifier": {
                                        "description": "YouTube URL or Textbook ISBN",
                                        "type": "string",
                                        "example": "0201896834"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Optional manual title for YouTube videos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resource retrieved or created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/resources/{id}": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Get a specific learning resource",
                "operationId": "a15c5328f3e07ab3950a6ee8cebdbc08",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/reviews/{review_id}/like": {
            "post": {
                "tags": [
                    "Likes"
                ],
                "summary": "Toggle a like on a review",
                "operationId": "f38eed904d92544ddd95185f13802162",
                "parameters": [
                    {
                        "name": "review_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Like toggled successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reports": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Report a resource or review to admins",
                "operationId": "318de4ef8123b483320dc9dbb45b6a43",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reason"
                                ],
                                "properties": {
                                    "learning_resource_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "review_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Invalid YouTube link"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Report submitted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reviews": {
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Leave a review on a learning resource",
                "operationId": "154c71f5f72b459d998f5ea451bd4c96",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "learning_resource_id",
                                    "rating",
                                    "content"
                                ],
                                "properties": {
                                    "learning_resource_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rating": {
                                        "description": "Must be between 1 and 5",
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Great explanation of integrals!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review saved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Resources",
            "description": "Resources"
        },
        {
            "name": "Likes",
            "description": "Likes"
        },
        {
            "name": "Reports",
            "description": "Reports"
        },
        {
            "name": "Reviews",
            "description": "Reviews"
        }
    ]
}